requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
json-rpc==1.15.0
Markdown==3.7
MarkupSafe==3.0.2
munch==4.0.0
packaging==25.0
paramiko==3.5.0
promise==2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.3.0
Werkzeug==3.1.3
